+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/machine/ModuleTechMachine.java
@@ -62,7 +62,8 @@ public class ModuleTechMachine
         "ZenBrickSawmill",
         "ZenBrickCrucible",
         "ZenBrickOven",
-        "ZenMechanicalCompactingBin"
+        "ZenMechanicalCompactingBin",
+        "ZenMechanicalHopper"
     };
 
     for (String plugin : craftTweakerPlugins) {